requirements.txt
asgiref==3.5.2
certifi==2022.9.24
charset-normalizer==2.1.1
Django==4.1.2
django-ckeditor==6.5.1
django-js-asset==2.0.0
idna==3.4
Pillow==9.2.0
python-dotenv==0.21.0
requests==2.28.1
sqlparse==0.4.3
urllib3==1.26.12
